backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dill==0.3.8
Django==5.0.8
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
isort==5.13.2
lxml==5.2.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
platformdirs==4.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.2.6
PyMySQL==1.1.1
requests==2.32.3
rich==13.7.1
ruff==0.6.3
sqlparse==0.5.0
tomlkit==0.13.2
urllib3==2.2.2
